Fix file names to match class names: prompt.py for Prompt, prompt_version.py for PromptVersion
