+++ mlflow/entities/model_registry/__init__.py
@@ -4,8 +4,8 @@ from mlflow.entities.model_registry.model_version_deployment_job_state import (
 )
 from mlflow.entities.model_registry.model_version_search import ModelVersionSearch
 from mlflow.entities.model_registry.model_version_tag import ModelVersionTag
-from mlflow.entities.model_registry.prompt import PromptVersion
-from mlflow.entities.model_registry.prompt_metadata import Prompt
+from mlflow.entities.model_registry.prompt import Prompt
+from mlflow.entities.model_registry.prompt_version import PromptVersion
 from mlflow.entities.model_registry.registered_model import RegisteredModel
 from mlflow.entities.model_registry.registered_model_alias import RegisteredModelAlias
 from mlflow.entities.model_registry.registered_model_deployment_job_state import (
